// Frontend/src/app/models/Search.ts
export class Search {
  ano: number = -1;
  tipo: string = "";
  regiao: string = "";
  estado: string = "";
  opcao: string = "";

  // constructor(ano: String, tipo: String, regiao: String, estado: String, opcao: String){
  //   this.ano = ano;
  //   this.tipo = tipo;
  //   this.regiao = regiao;
  //   this.estado = estado;
  //   this.opcao = opcao;
  // }
}